# server/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"error": {
						"additionalProperties": false,
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"errorStack": {}
				},
				"required": [
					"success",
					"message",
					"error",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegisterPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"repeatPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"repeatPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOTReqPayload": {
				"properties": {
					"batteryVoltage": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"humidity": {
						"type": "number",
						"format": "double"
					},
					"soilMoisture": {
						"type": "number",
						"format": "double"
					},
					"isCharging": {
						"type": "boolean"
					},
					"time": {
						"type": "string"
					}
				},
				"required": [
					"batteryVoltage",
					"temperature",
					"humidity",
					"soilMoisture",
					"isCharging",
					"time"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "soil-monitoring-api",
		"version": "1.0.0",
		"description": "An api server for soil monitoring system focused on robust security and performance in IOT applications",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/user/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					}
				},
				"description": "Used for user registratiion.\r\nSupply the user's name, unique email and password",
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user logged in successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user logged out successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/user/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "password changed successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/user/get-all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "users fetched successfully",
											"data": [
												{
													"email": "john@doe.com",
													"name": "john doe",
													"createdAt": "2021-09-15T17:07:15.000Z",
													"updatedAt": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true,
												false
											]
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "auth",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/iot/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"mac": "00:00:00:00:00:00"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"IOT"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"mac": {
										"type": "string"
									}
								},
								"required": [
									"mac"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/iot/new-token": {
			"post": {
				"operationId": "NewToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"totp": "123456",
											"mac": "00:00:00:00:00:00"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"IOT"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-package",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"mac": {
										"type": "string"
									}
								},
								"required": [
									"mac"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/iot/data": {
			"post": {
				"operationId": "Data",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"IOT"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-forwarded-for",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "X-TOTP",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-package",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/IOTReqPayload"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetData",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"IOT"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"csv",
								"json"
							]
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}